requirements.txt
asgiref==3.4.1
boto3==1.18.32
botocore==1.21.32
dj-database-url==0.5.0
Django==3.2.6
django-filter==2.4.0
django-heroku==0.3.1
django-rest-passwordreset==1.2.0
django-storages==1.11.1
djangorestframework==3.12.4
gunicorn==20.1.0
importlib-metadata==4.6.4
jmespath==0.10.0
Markdown==3.3.4
Pillow==8.3.1
psycopg2==2.9.1
python-dateutil==2.8.2
pytz==2021.1
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.6
whitenoise==5.3.0
zipp==3.5.0
